pyproject.toml
[tool.poetry]
name = "graphrag-agent"
version = "0.1.0"
description = "GraphRAG application with Gemini LLM and React frontend"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-multipart = "^0.0.9"
google-genai = "^1.41.0"
google-adk = "^1.16.0"
neo4j = "^5.14.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
aiofiles = "^23.2.1"
pypdf = "^3.17.0"
python-docx = "^1.1.0"
numpy = "^1.26.0"
sentence-transformers = "^2.2.2"
gitpython = "^3.1.40"
mcp = "^1.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
